+++ .gitignore
@@ -12,7 +12,7 @@
 /build
 
 #TODO
-TODO
+TODO.txt
 
 temp.js
