+++ src/component/Filter.js
@@ -1,14 +1,9 @@
 import React from 'react';
 
 const Filter = () => {
-
-  return (
-    <div className='py-4'>
-      {/* <form>
-        <input type='dropdown' />
-      </form> */}
-    </div>
-  );
-}
+  return <div className='py-4'>
+      
+    </div>;
+};
 
 export default Filter;
